requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
deepl==1.18.0
discord.py==2.3.2
emoji==2.12.1
frozenlist==1.4.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.5
pyaml==24.4.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.1
yarl==1.9.4
